Crate numeric_search

Run classic binary or exponential search on integer or floating point numbers.

Both classic binary seach and exponential search has their streangth and weakness.

The classic binary search is (about $\times 2$)) faster in worst case.

The classic binary search never fails, even if the partition function is not monotone.

We do not have to know the limit values to use the exponential search.

The running time of the exponential search is output-sensitively fast.

Traits

Float

Floating pont number.

Signed

Signed integers.

Unsigned

Unsigned integers.

Functions

binary_search_signed

Run a binary search search on signed numbers.

binary_search_unsigned

Run a binary search search on unsigned numbers.

exp_search_float

Run an exponential search on floating point numbers.

exp_search_signed

Run an exponential search on unsigned numbers.

exp_search_unsigned

Run an exponential search on unsigned numbers.