Function numeric_search::exp_search_float

pub fn exp_search_float<T: Float>(f: impl FnMut(T) -> bool) -> T

Run an exponential search on floating point numbers.

Given a binary function $f: \mathbb R \rightarrow \mathtt { bool }$, it trys to find a normal number $x \in \mathbb R$ satisfying $\neg f( \mathtt { prev } ( x ) ) \land f(x)$ (where $\mathtt { prev } ( x )$ is the previous normal number of $x$).

This trial always succeeds provided that

$f$ is monotone from $\mathbb { F }$ to $\mathbb { T }$ and

there exist such $x$’s in -T::MAX.sqrt()..=T::MAX.sqrt().

If it falis to find it, it returns T::{INFINITY, NEG_INFINITY}.

Examples

They are some usual usages where the function $f$ is monotone.

assert_eq!(exp_search_float(|x| 2.5 <= x), 2.5);
assert_eq!(exp_search_float(|x| -2.5 <= x), -2.5);