Function numeric_search::exp_search_unsigned

pub fn exp_search_unsigned<T: Unsigned>(f: impl FnMut(T) -> bool) -> Option<T>

Run an exponential search on unsigned numbers.

Given a function $f: \mathbb N \to \mathtt { bool }$, it try to find $x \in \mathbb N$ satisfying $\neg f ( x - 1 ) \land f ( x )$ (where we assume that $\neg f ( -1 )$).

This trial always succeeds provided that

$f$ is monotone from $\mathbb { F }$ to $\mathbb { T }$ and

$f$ is not always-true

If it falis to find it, it returns None.

Examples

They are some usual usages where the function $f$ is monotone.

assert_eq!(exp_search_unsigned(|x: u32| 6 <= x), Some(6));
assert_eq!(exp_search_unsigned(|_: u32| false), None);
assert_eq!(exp_search_unsigned(|_: u32| true), Some(0));